+++ lib/ui/models/ui_theme_model.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+// Model for storing (and dynamically updating) theme of app
+class MyThemeModel with ChangeNotifier {
+  bool _isDarkTheme = false;
+
+  bool get isDarkTheme => _isDarkTheme;
+
+  void changeTheme() {
+    _isDarkTheme = !_isDarkTheme;
+    notifyListeners();
+  }
+}
