+++ pubspec.yaml
@@ -21,6 +21,8 @@ flutter:
   uses-material-design: true
   assets:
     - res/
+    - res/images/
+
   fonts:
     - family: Roboto
       fonts:
